add ParticipantDao tests

+++ participants/src/Piipan.Participants/Piipan.Participants.Core/DataAccessObjects/IParticipantDao.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Piipan.Participants.Core.Models;
@@ -6,7 +7,7 @@ namespace Piipan.Participants.Core.DataAccessObjects
 {
     public interface IParticipantDao
     {
-        Task<IEnumerable<ParticipantDbo>> GetParticipants(string ldsHash, int uploadId);
+        Task<IEnumerable<ParticipantDbo>> GetParticipants(string ldsHash, Int64 uploadId);
         Task AddParticipants(IEnumerable<ParticipantDbo> participants);
     }
 }
